move dependencies to data, make []Command as Commands, change replicaId's type to uint8, make everything in data packge exportable

+++ replica/replica.go
@@ -25,15 +25,15 @@ const (
 // ****************************
 
 type Replica struct {
-	Id             int
-	Size           int
+	Id             uint8
+	Size           uint8
 	MaxInstanceNum []uint64 // the highest instance number seen for each replica
 	InstanceMatrix [][]*Instance
 	StateMachine   epaxos.StateMachine
 	Epoch          uint32
 }
 
-func New(replicaId, size int, sm epaxos.StateMachine) (r *Replica) {
+func New(replicaId, size uint8, sm epaxos.StateMachine) (r *Replica) {
 	r = &Replica{
 		Id:             replicaId,
 		Size:           size,
@@ -43,7 +43,7 @@ func New(replicaId, size int, sm epaxos.StateMachine) (r *Replica) {
 		Epoch:          0,
 	}
 
-	for i := 0; i < size; i++ {
+	for i := uint8(0); i < size; i++ {
 		r.InstanceMatrix[i] = make([]*Instance, 1024)
 		r.MaxInstanceNum[i] = conflictNotFound
 	}
